[KVConnector][Mooncake] Keep layer index helper local
