--- vllm/utils/layer_utils.py
@@ -1,35 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-
-
-def extract_layer_index(layer_name: str, num_attn_module: int = 1) -> int:
-    """
-    Extract the layer index from the module name.
-    Examples:
-    - "encoder.layers.0" -> 0
-    - "encoder.layers.1.self_attn" -> 1
-    - "2.self_attn" -> 2
-    - "model.encoder.layers.0.sub.1" -> ValueError if num_attn_module == 1
-    """
-    subnames = layer_name.split(".")
-    int_vals: list[int] = []
-    for subname in subnames:
-        try:
-            int_vals.append(int(subname))
-        except ValueError:
-            continue
-    if num_attn_module == 1 or "attn" not in layer_name:
-        assert len(int_vals) == 1, (
-            f"layer name {layer_name} should only contain one integer"
-        )
-
-        return int_vals[0]
-
-    assert len(int_vals) <= 2, (
-        f"layer name {layer_name} should contain most two integers"
-    )
-    return (
-        int_vals[0] * num_attn_module + int_vals[1]
-        if len(int_vals) == 2
-        else int_vals[0]
-    )
